Revert "Merge pull request #1 from MobikTech/Lab_3"
This reverts commit 7122b7476dc7c906e125db7637195a4661e9c527, reversing
changes made to 7e248453f5475b50512e6f7097033a48464e2938.

+++ FinanceManager.DAL/Entities/Account.cs
@@ -1,4 +1,5 @@
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
 
 namespace FinanceManager.DAL.Entities
 {
@@ -7,7 +8,7 @@ namespace FinanceManager.DAL.Entities
         public int Id { get; set; }
         public string Number { get; set; }
         
-        public decimal Count { get; set; }
+        public int Count { get; set; }
         
         public ICollection<Transaction> TransactionsAsSource { get; set; }
